+++ pydsm/__init__.py
@@ -4,5 +4,5 @@
 
 __author__ = """Nicky Sandhu"""
 __email__ = 'psandhu@water.ca.gov'
-__version__ = '0.1.8'
+__version__ = '0.2.0'
 __all__=[]