+++ pydsm/input/__init__.py
@@ -1,0 +1,1 @@
+## Module to read/write DSM2 standard input as pandas data frames